requirements.txt
# Embeddings + vector search
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4

# API + config
fastapi==0.114.0
uvicorn==0.30.0
pydantic==2.8.2
python-dotenv==1.0.1

# Data handling
pandas==2.2.2
pypdf==4.3.1
pyarrow==17.0.0        

neo4j==5.22.0
